extract GalileoHandler to its own file and follow styling of .env.example of other samples
