+++ python/chatbot/elevenlabs-chatbot/.env.example
@@ -1,9 +1,7 @@
-# ElevenLabs Configuration
-ELEVENLABS_API_KEY=your-elevenlabs-api-key
-ELEVENLABS_AGENT_ID=your-agent-id
-
-# Galileo Configuration (for local instance)
+# Galileo Configuration
 GALILEO_API_KEY=your-galileo-api-key
-GALILEO_CONSOLE_URL=your-galileo-console-url
 GALILEO_PROJECT_NAME=elevenlabs-voice-poc
 GALILEO_LOG_STREAM=voice-chatbot
+
+# Provide the console url below if you are using a custom deployment, and not using app.galileo.ai
+# GALILEO_CONSOLE_URL=your-galileo-console-url   # Optional if you are using a hosted version of Galileo
